+++ handlers/otp.go
@@ -6,16 +6,15 @@ package handlers
 
 import (
 	"bytes"
+	"encoding/base64"
 	"image/png"
 	"net/http"
-	"strconv"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/justinas/nosurf"
 	"github.com/pquerna/otp"
 	"github.com/ubccr/goipa"
 	"github.com/ubccr/mokey/app"
-	"github.com/ubccr/mokey/model"
 )
 
 func TwoFactorHandler(ctx *app.AppContext) http.Handler {
@@ -35,34 +34,65 @@ func TwoFactorHandler(ctx *app.AppContext) http.Handler {
 		message := ""
 
 		if r.Method == "POST" {
-			sid := session.Values[app.CookieKeySID]
+			// These operations require admin privs. TODO: should we make this
+			// configurable?
+			c := app.NewIpaClient(true)
 			action := r.FormValue("action")
 			if action == "remove" {
-				err = DisableTOTP(ctx, string(user.Uid), sid.(string))
-
+				// Remove any auth types which will fall back to FreeIPA global default.
+				err := c.SetAuthTypes(string(user.Uid), nil)
 				if err == nil {
-					message = "TOTP Disabled"
+					user.AuthTypes = []string{}
 				} else {
+					log.WithFields(log.Fields{
+						"user":  string(user.Uid),
+						"error": err,
+					}).Error("failed to reset auth types to default")
 					message = "Failed to disable TOTP. Please contact your administrator"
+				}
+			} else if action == "enable" {
+				err := c.SetAuthTypes(string(user.Uid), []string{"otp"})
+				if err == nil {
+					user.AuthTypes = []string{"otp"}
+				} else {
+					log.WithFields(log.Fields{
+						"user":  string(user.Uid),
+						"error": err,
+					}).Error("failed to set auth types to otp")
+					message = "Failed to enable TOTP. Please contact your administrator"
+				}
+
+				sid := session.Values[app.CookieKeySID]
+				c.SetSession(sid.(string))
+				tokens, err := c.FetchOTPTokens(string(user.Uid))
+				if err != nil {
+					log.WithFields(log.Fields{
+						"user":  string(user.Uid),
+						"error": err,
+					}).Error("failed to fetch OTP tokens")
+					ctx.RenderError(w, http.StatusInternalServerError)
+					return
+				}
+				if len(tokens) == 0 {
+					addNewToken(ctx, w, r)
+					return
 				}
 			}
 		}
 
-		token, _ := model.FetchConfirmedOTPToken(ctx.Db, string(user.Uid))
-
 		vars := map[string]interface{}{
-			"flashes":   session.Flashes(),
-			"totptoken": token,
-			"token":     nosurf.Token(r),
-			"message":   message,
-			"user":      user}
+			"flashes":    session.Flashes(),
+			"otpenabled": user.OTPOnly(),
+			"token":      nosurf.Token(r),
+			"message":    message,
+			"user":       user}
 
 		session.Save(r, w)
 		ctx.RenderTemplate(w, "2fa.html", vars)
 	})
 }
 
-func EnableTOTPHandler(ctx *app.AppContext) http.Handler {
+func OTPTokensHandler(ctx *app.AppContext) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		user := ctx.GetUser(r)
 		if user == nil {
@@ -73,21 +103,6 @@ func EnableTOTPHandler(ctx *app.AppContext) http.Handler {
 		session, err := ctx.GetSession(r)
 		if err != nil {
 			ctx.RenderError(w, http.StatusInternalServerError)
-			return
-		}
-
-		_, err = model.FetchConfirmedOTPToken(ctx.Db, string(user.Uid))
-		if err == nil {
-			session.AddFlash("TOTP Already Enabled")
-			session.Save(r, w)
-			http.Redirect(w, r, "/2fa", 302)
-			return
-		}
-
-		_, err = model.FetchUnconfirmedOTPToken(ctx.Db, string(user.Uid))
-		if err == nil {
-			log.Info("User already has an uncofirmed TOTP")
-			http.Redirect(w, r, "/2fa/verify", 302)
 			return
 		}
 
@@ -95,170 +110,151 @@ func EnableTOTPHandler(ctx *app.AppContext) http.Handler {
 		c := app.NewIpaClient(false)
 		c.SetSession(sid.(string))
 
-		uri, err := c.AddTOTPToken(string(user.Uid), ipa.AlgorithmSHA1, ipa.DigitsSix, 30)
-		if err != nil {
-			if ierr, ok := err.(*ipa.IpaError); ok && ierr.Code == 4002 {
-				// 4002 user already has a TOTP that mokey doesn't know about
-				session.AddFlash("You already have a TOTP set in FreeIPA. Please contact your administrator or remove existing TOTP manually.")
-				session.Save(r, w)
-				http.Redirect(w, r, "/2fa", 302)
-				return
-			} else {
-				log.WithFields(log.Fields{
-					"user":  string(user.Uid),
-					"error": err,
-				}).Error("Failed to create TOTP")
-				ctx.RenderError(w, http.StatusInternalServerError)
-				return
-			}
-		}
-
-		err = model.StoreOTPToken(ctx.Db, string(user.Uid), uri)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"user":  string(user.Uid),
-				"error": err,
-			}).Error("Failed to save TOTP to database")
-			ctx.RenderError(w, http.StatusInternalServerError)
-			return
-		}
-
-		http.Redirect(w, r, "/2fa/verify", 302)
-		return
-	})
-}
-
-func VerifyTOTPHandler(ctx *app.AppContext) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		user := ctx.GetUser(r)
-		if user == nil {
-			ctx.RenderError(w, http.StatusInternalServerError)
-			return
-		}
-
-		session, err := ctx.GetSession(r)
-		if err != nil {
-			ctx.RenderError(w, http.StatusInternalServerError)
-			return
-		}
-
-		_, err = model.FetchConfirmedOTPToken(ctx.Db, string(user.Uid))
-		if err == nil {
-			session.AddFlash("TOTP Already Enabled")
-			session.Save(r, w)
-			http.Redirect(w, r, "/2fa", 302)
-			return
-		}
-
-		token, err := model.FetchUnconfirmedOTPToken(ctx.Db, string(user.Uid))
-		if err != nil {
-			log.WithFields(log.Fields{
-				"user":  string(user.Uid),
-				"error": err,
-			}).Error("Failed to fetch TOTP")
-			ctx.RenderNotFound(w)
-			return
-		}
-
 		message := ""
+
 		if r.Method == "POST" {
-			code := r.FormValue("code")
-			if token.Validate(code) {
-				err := model.ConfirmOTPToken(ctx.Db, string(user.Uid))
+			action := r.FormValue("action")
+			uuid := r.FormValue("uuid")
+			log.WithFields(log.Fields{
+				"user":   string(user.Uid),
+				"uuid":   uuid,
+				"action": action,
+			}).Info("otptokens action")
+
+			if action == "delete" && len(uuid) > 0 {
+				err := c.RemoveOTPToken(uuid)
 				if err != nil {
 					log.WithFields(log.Fields{
 						"user":  string(user.Uid),
+						"uuid":  uuid,
 						"error": err,
-					}).Error("Failed to confirm TOTP")
-					ctx.RenderError(w, http.StatusInternalServerError)
-					return
+					}).Error("failed to remove OTP Token")
+					message = "Failed to remove OTP Token"
 				}
-
-				session.AddFlash("TOTP Enabled Successfully")
-				session.Save(r, w)
-				http.Redirect(w, r, "/2fa", 302)
+			} else if action == "enable" && len(uuid) > 0 {
+				err := c.EnableOTPToken(uuid)
+				if err != nil {
+					log.WithFields(log.Fields{
+						"user":  string(user.Uid),
+						"uuid":  uuid,
+						"error": err,
+					}).Error("failed to enable OTP Token")
+					message = "Failed to enable OTP Token"
+				}
+			} else if action == "disable" && len(uuid) > 0 {
+				err := c.DisableOTPToken(uuid)
+				if err != nil {
+					log.WithFields(log.Fields{
+						"user":  string(user.Uid),
+						"uuid":  uuid,
+						"error": err,
+					}).Error("failed to disable OTP Token")
+					message = "Failed to disable OTP Token"
+				}
+			} else if action == "add" {
+				addNewToken(ctx, w, r)
 				return
 			}
+		}
 
-			message = "Invalid code. Try again"
+		tokens, err := c.FetchOTPTokens(string(user.Uid))
+		if err != nil {
+			ctx.RenderError(w, http.StatusInternalServerError)
+			return
 		}
 
 		vars := map[string]interface{}{
-			"token":   nosurf.Token(r),
-			"message": message,
-			"user":    user}
+			"flashes":   session.Flashes(),
+			"token":     nosurf.Token(r),
+			"message":   message,
+			"otptokens": tokens,
+			"user":      user}
 
-		ctx.RenderTemplate(w, "verify-totp.html", vars)
+		session.Save(r, w)
+		ctx.RenderTemplate(w, "otp-tokens.html", vars)
 	})
 }
 
-func DisableTOTP(ctx *app.AppContext, uid, sid string) error {
+func addNewToken(ctx *app.AppContext, w http.ResponseWriter, r *http.Request) {
+	user := ctx.GetUser(r)
+	if user == nil {
+		ctx.RenderError(w, http.StatusInternalServerError)
+		return
+	}
+
+	session, err := ctx.GetSession(r)
+	if err != nil {
+		ctx.RenderError(w, http.StatusInternalServerError)
+		return
+	}
+
+	sid := session.Values[app.CookieKeySID]
+	c := app.NewIpaClient(false)
+	c.SetSession(sid.(string))
+
+	otptoken, err := c.AddTOTPToken(string(user.Uid), ipa.AlgorithmSHA1, ipa.DigitsSix, 30)
+	if err != nil {
+		log.WithFields(log.Fields{
+			"user":  string(user.Uid),
+			"error": err,
+		}).Error("Failed to create TOTP")
+		ctx.RenderError(w, http.StatusInternalServerError)
+		return
+	}
+
+	otpdata, err := QRCode(otptoken)
+	if err != nil {
+		log.WithFields(log.Fields{
+			"user":  string(user.Uid),
+			"error": err,
+		}).Error("failed to render TOTP token as QRCode image")
+		ctx.RenderError(w, http.StatusInternalServerError)
+		return
+	}
+
+	vars := map[string]interface{}{
+		"otpdata":  otpdata,
+		"otptoken": otptoken,
+		"user":     user}
+
+	ctx.RenderTemplate(w, "verify-totp.html", vars)
+}
+
+func RemoveAllOTPTokens(uid, sid string) error {
 	c := app.NewIpaClient(false)
 	c.SetSession(sid)
 
-	err := c.RemoveOTPToken(uid)
+	tokens, err := c.FetchOTPTokens(uid)
 	if err != nil {
-		if ierr, ok := err.(*ipa.IpaError); ok {
-			// 4001 not found means user didn't have a token so we ignore
-			if ierr.Code != 4001 {
-				log.WithFields(log.Fields{
-					"user":  uid,
-					"error": err,
-				}).Error("Failed to remove TOTP from FreeIPA")
-				return err
-			}
-		} else {
-			log.WithFields(log.Fields{
-				"user":  uid,
-				"error": err,
-			}).Error("Failed to remove TOTP from FreeIPA")
-			return err
-		}
+		return err
 	}
 
-	err = model.RemoveOTPToken(ctx.Db, uid)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"user":  uid,
-			"error": err,
-		}).Error("Failed to remove TOTP")
-		return err
+	for _, t := range tokens {
+		err = c.RemoveOTPToken(string(t.UUID))
+		if err != nil {
+			return err
+		}
 	}
 
 	return nil
 }
 
-func QRCodeHandler(ctx *app.AppContext) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		user := ctx.GetUser(r)
-		if user == nil {
-			ctx.RenderError(w, http.StatusInternalServerError)
-			return
-		}
+func QRCode(otptoken *ipa.OTPToken) (string, error) {
+	if otptoken == nil {
+		return "", nil
+	}
 
-		token, err := model.FetchUnconfirmedOTPToken(ctx.Db, string(user.Uid))
-		if err != nil {
-			w.WriteHeader(http.StatusNotFound)
-			return
-		}
+	key, err := otp.NewKeyFromURL(otptoken.URI)
+	if err != nil {
+		return "", err
+	}
 
-		key, err := otp.NewKeyFromURL(token.URI)
-		if err != nil {
-			w.WriteHeader(http.StatusNotFound)
-			return
-		}
+	var buf bytes.Buffer
+	img, err := key.Image(250, 250)
+	if err != nil {
+		return "", err
+	}
 
-		var buf bytes.Buffer
-		img, err := key.Image(250, 250)
-		if err != nil {
-			w.WriteHeader(http.StatusNotFound)
-			return
-		}
-		png.Encode(&buf, img)
-
-		w.Header().Set("Content-Type", "image/png")
-		w.Header().Set("Content-Length", strconv.Itoa(len(buf.Bytes())))
-
-		buf.WriteTo(w)
-	})
+	png.Encode(&buf, img)
+	return base64.StdEncoding.EncodeToString(buf.Bytes()), nil
 }