+++ model/db.go
@@ -1,0 +1,24 @@
+// Copyright 2015 mokey Authors. All rights reserved.
+// Use of this source code is governed by a BSD style
+// license that can be found in the LICENSE file.
+
+package model
+
+import (
+	_ "github.com/go-sql-driver/mysql"
+	"github.com/jmoiron/sqlx"
+)
+
+func NewDB(driver, dsn string) (*sqlx.DB, error) {
+	db, err := sqlx.Open(driver, dsn)
+	if err != nil {
+		return nil, err
+	}
+
+	err = db.Ping()
+	if err != nil {
+		return nil, err
+	}
+
+	return db, nil
+}